refactor(coordinator): narrow public surface to driver shape (#553)
C9: the package barrel now exports exactly the Execution.Driver-shaped
delivery core — createWorkerManager plus the six-method WorkerManager
type (deliver/cancel/send/stats/waitUntilReady/shutdown) and its
parameter/result types (DeliverTask, WorkerManagerStats, port
param/result types InboundWait*/ToolCall*).

- WorkerManagerConfig and WorkerPorts fold into factory-private config:
  callers pass literals; slotWaitTimeoutMs/maxQueuedDeliveries remain
  pool-internal knobs reachable only from the coordinator's own tests.
- killWorker leaves the public WorkerManager type; the crash-recovery
  test reaches the chaos hook via killWorkerForTest on the internal
  worker-manager barrel (instanceof-guarded, no casts).
- ToolCallCancelParams drops out of the re-export chain end to end.
- worker-pool.ts stops re-exporting types, resolving the two
  grandfathered DeliverTask dead-export entries (baseline 21 -> 19).
- facade-removal test now pins the runtime export list to exactly
  ["createWorkerManager"].

Consumers re-mapped at head: apps/server worker-full-stack fake drops
killWorker; all other sites compile unchanged. No as-bridges, no
pass-through wrappers.
