+++ packages/coordinator/src/index.ts
@@ -1,13 +1,15 @@
+// Public surface (#553 C9): the Execution.Driver-shaped delivery core —
+// deliver/cancel/send/stats/waitUntilReady/shutdown — plus its parameter and
+// result types. Pool construction knobs (`WorkerManagerConfig`/`WorkerPorts`)
+// are factory-private; worker-supervision never leaves the package.
 export { createWorkerManager } from "./worker-manager";
 export type {
+  DeliverTask,
   InboundWaitParams,
   InboundWaitResult,
-  ToolCallCancelParams,
   ToolCallContext,
   ToolCallParams,
   ToolCallResult,
   WorkerManager,
-  WorkerManagerConfig,
   WorkerManagerStats,
-  WorkerPorts,
 } from "./worker-manager";
