+++ packages/coordinator/src/worker-supervision/supervisor.ts
@@ -27,7 +27,6 @@ const WORKER_CONNECT_TIMEOUT_MS = 10_000;
 export type {
   InboundWaitParams,
   InboundWaitResult,
-  ToolCallCancelParams,
   ToolCallContext,
   ToolCallParams,
   ToolCallResult,
